requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# MCP SDK
mcp==0.9.0

# Microsoft Agent Framework
azure-ai-agent==0.1.0

# AI/LLM clients
openai==1.3.0
anthropic==0.7.0

# Database clients
pymongo==4.6.0
redis==5.0.1
motor==3.3.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
